requirements.txt
absl-py==2.1.0
alabaster==0.7.16
ale-py==0.8.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
baselines==0.1.5
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==1.2.2
cmake==3.25.0
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
decorator==5.1.1
dill==0.3.8
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2023.4.0
future==1.0.0
gast==0.5.4
glfw==2.7.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.62.1
gtimer==1.0.0b5
gym==0.15.7
gym-notices==0.0.8
h5py==3.11.0
idna @ file:///croot/idna_1666125576474/work
ImageHash==4.3.1
imageio==2.34.0
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
intel-openmp==2024.1.0
ipdb==0.13.13
ipython==8.18.1
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
keras==3.2.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lit==15.0.7
lockfile==0.12.2
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2024.1.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.2.0
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath @ file:///croot/mpmath_1690848262763/work
mujoco==2.2.0
mujoco-py==2.0.2.5
namex==0.0.8
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
numpydoc==1.7.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
olefile==0.47
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1707233021655/work
pluggy==1.4.0
progressbar2==4.4.2
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
pygame==2.1.0
pyglet==1.5.0
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytest==8.1.1
pytest-instafail==0.5.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.0.0
requests @ file:///croot/requests_1707355572290/work
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setproctitle==1.3.3
sip==6.8.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
statsmodels==0.14.1
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
tomli==2.0.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1707770551213/work
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
zmq==0.0.0
